// client/tooling/uml-language-tooling/src/types/multiplicity.ts
/**********************************************************************************
 * Copyright (c) 2025 borkdominik and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the MIT License which is available at https://opensource.org/licenses/MIT.
 *
 * SPDX-License-Identifier: MIT
 **********************************************************************************/
export enum Multiplicity {
  ZERO_TO_N = "*",
  ONE_TO_N = "+",
  ONE_TO_ONE = "1",
}